backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-jose[cryptography]
pydantic
pydantic-settings
